+++ client-js/app/scripts/app.ts
@@ -4,6 +4,8 @@ import page = require("page");
 
 import registry = require("./registry");
 import datavault = require("./datavault");
+import dumb_server = require("./dumb_server");
+import client_rpc = require("./client_rpc");
 
 // autobinding template which is the main ui container
 var app: any = document.querySelector('#app');
@@ -23,9 +25,9 @@ window.addEventListener('WebComponentsReady', function() {
     page(e.detail.path);
   });
 
-  var socket = new registry.JsonRpcSocket('ws://localhost:9000/api/socket2');
+  var socket = new client_rpc.JsonRpcSocket('ws://localhost:9000/api/socket2');
   var reg = new registry.RegistryServiceJsonRpc(socket);
-  var dumb = new registry.DumbServiceJsonRpc(socket);
+  var dumb = new dumb_server.DumbServiceJsonRpc(socket);
 
   //var reg = new registry.RegistryService('http://localhost:9000');
   var dv = new datavault.DataVaultService('http://localhost:9000');
@@ -43,7 +45,7 @@ window.addEventListener('WebComponentsReady', function() {
 
   function loadRegistry(path: Array<string>) {
     console.log('loading registry:', path);
-    dumb.dumb_echo({inp: "lalala"});
+    dumb.dumb_echo({inp: "message_2045"});
     reg.dir({path: path}).then((listing) => {
       console.log(listing);
